backend/requirements.txt
asgiref==3.8.1
astroid==3.3.8
black==24.10.0
click==8.1.8
coverage==7.6.10
dill==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.1
pylint==3.3.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
whitenoise==6.8.2
